+++ RELEASE_NOTES.md
@@ -3,3 +3,5 @@ v0.84.0
 - keytrans: Verify signatures from all auditors
 
 - Java: Fixed `IdentityKeyPair(byte[])` to correctly declare that it throws InvalidKeyException.
+
+- Node: Add BackupsJsonExporter, to convert Backup proto objects to human-readable JSON for export.